backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# LangChain e RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2

# Processamento de texto
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilitários
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1 